refactor: inject dependence TOKEN to desacople modules
