+++ src/_test_/entity/user.spec.ts
@@ -1,13 +1,15 @@
 import User from '../../entity/User';
 import MockFalseRepo from './MockFalseRepo';
-import MockTrueRepo from './MockTrueRepo';
+import MockToken from './MockToken';
+import MockTrueRepo from './MockTrueRepo copy';
 
 const mock = new MockTrueRepo();
 const mockFalse = new MockFalseRepo();
+const mockToken = new MockToken();
 
 describe('User', () => {
   it('should return true if user already exists', async () => {
-    const user = new User(mock);
+    const user = new User(mock, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.exist();
@@ -15,7 +17,7 @@ describe('User', () => {
   });
 
   it('should return false to create if user already exists', async () => {
-    const user = new User(mock);
+    const user = new User(mock, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.create();
@@ -23,7 +25,7 @@ describe('User', () => {
   });
 
   it('should return true to update if user already exists', async () => {
-    const user = new User(mock);
+    const user = new User(mock, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.update(user.id);
@@ -31,7 +33,7 @@ describe('User', () => {
   });
 
   it('should return true to delete if user already exists', async () => {
-    const user = new User(mock);
+    const user = new User(mock, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.delete(user.id);
@@ -39,7 +41,7 @@ describe('User', () => {
   });
 
   it('should return true to verify if id user already exists', async () => {
-    const user = new User(mock);
+    const user = new User(mock, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.verifyExistById(user.id);
@@ -47,7 +49,7 @@ describe('User', () => {
   });
 
   it('should return object User when have a user with this id', async () => {
-    const user = new User(mock);
+    const user = new User(mock, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const result = await user.getById(user.id);
@@ -58,7 +60,7 @@ describe('User', () => {
   });
 
   it('should return false if user already exists', async () => {
-    const user = new User(mockFalse);
+    const user = new User(mockFalse, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.exist();
@@ -66,7 +68,7 @@ describe('User', () => {
   });
 
   it('should return true to create if user does not exists', async () => {
-    const user = new User(mockFalse);
+    const user = new User(mockFalse, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.create();
@@ -74,7 +76,7 @@ describe('User', () => {
   });
 
   it('should return false to update if user does not exists', async () => {
-    const user = new User(mockFalse);
+    const user = new User(mockFalse, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.update(user.id);
@@ -82,7 +84,7 @@ describe('User', () => {
   });
 
   it('should return false to delete if user does not exists', async () => {
-    const user = new User(mockFalse);
+    const user = new User(mockFalse, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.delete(user.id);
@@ -90,7 +92,7 @@ describe('User', () => {
   });
 
   it('should return false to verify, if id user does not exists on DB', async () => {
-    const user = new User(mockFalse);
+    const user = new User(mockFalse, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const exists = await user.verifyExistById(user.id);
@@ -98,7 +100,7 @@ describe('User', () => {
   });
 
   it('should return false when have no user with this id', async () => {
-    const user = new User(mockFalse);
+    const user = new User(mockFalse, mockToken);
     user.id = 'ad163ac6-4fa8-11ec-81d3-0242ac130003';
 
     const result = await user.getById(user.id);
