+++ src/infrastructure/token/IToken.ts
@@ -1,0 +1,8 @@
+import User from '../../entity/User';
+import dtoToken from './dtoToken';
+
+interface IToken {
+  generate(user: User): dtoToken;
+}
+
+export default IToken;
